+++ Shading/components/com_camera.ts
@@ -1,27 +1,1 @@
-import {create} from "../../common/mat4.js";
-import {Mat4, Vec3} from "../../common/math.js";
-import {Entity, Game} from "../game.js";
-import {Has} from "../world.js";
-
-export interface Camera {
-    FovY: number;
-    Near: number;
-    Far: number;
-    Projection: Mat4;
-    Position: Vec3;
-    Pv: Mat4;
-}
-
-export function camera(fovy: number, near: number, far: number) {
-    return (game: Game, entity: Entity) => {
-        game.World.Signature[entity] |= Has.Camera;
-        game.World.Camera[entity] = {
-            FovY: fovy,
-            Near: near,
-            Far: far,
-            Projection: create(),
-            Position: [0, 0, 0],
-            Pv: create(),
-        };
-    };
-}
+../../core/components/com_camera.ts
