+++ Shading/components/com_rotate.ts
@@ -1,17 +1,1 @@
-import {Vec3} from "../../common/math.js";
-import {Entity, Game} from "../game.js";
-import {Has} from "../world.js";
-
-export interface Rotate {
-    /** Eulers per second on each axis */
-    Rotation: Vec3;
-}
-
-export function rotate(rotation: Vec3) {
-    return (game: Game, entity: Entity) => {
-        game.World.Signature[entity] |= Has.Rotate;
-        game.World.Rotate[entity] = {
-            Rotation: rotation,
-        };
-    };
-}
+../../core/components/com_rotate.ts
