+++ Shading/scenes/sce_stage.ts
@@ -1,8 +1,6 @@
 import {blueprint_camera} from "../blueprints/blu_camera.js";
 import {light_directional, light_point} from "../components/com_light.js";
-import {render_basic} from "../components/com_render_basic.js";
-import {render_diffuse} from "../components/com_render_diffuse.js";
-import {render_specular} from "../components/com_render_specular.js";
+import {render_basic, render_diffuse, render_specular} from "../components/com_render1.js";
 import {rotate} from "../components/com_rotate.js";
 import {instantiate} from "../core.js";
 import {Game} from "../game.js";
